backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Google API dependencies (optimized versions)
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
google-generativeai>=0.3.0

# Core utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
httpx>=0.25.2

# AI/ML dependencies (optimized for space)
chromadb>=0.4.22
sentence-transformers>=2.2.2
tiktoken>=0.5.2

# Document processing (lightweight alternatives)
pypdf>=3.17.1
python-docx>=1.1.0

# Security & Rate Limiting
slowapi>=0.1.9

# Monitoring & Observability
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.2

# Code Quality (optional, for development)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0